Swagger

+++ controle-frequencia/src/main/java/com/senai/backend/controle_frequencia/repositories/TurmaRepository.java
@@ -5,9 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TurmaRepository extends JpaRepository<Turma, Long>{
+public interface TurmaRepository extends JpaRepository<Turma, Integer>{
 
-}
- 
-    
-
+}
